// web/src/model/sportradar/models/GameSummary/Interfaces/GameSummaryPlayer.ts
export interface Period {
    type: string;
    id: string;
    number: number;
    sequence: number;
    minutes: string;
    field_goals_made: number;
    field_goals_att: number;
    field_goals_pct: number;
    three_points_made: number;
    three_points_att: number;
    three_points_pct: number;
    two_points_made: number;
    two_points_att: number;
    two_points_pct: number;
    blocked_att: number;
    free_throws_made: number;
    free_throws_att: number;
    free_throws_pct: number;
    offensive_rebounds: number;
    defensive_rebounds: number;
    rebounds: number;
    assists: number;
    turnovers: number;
    steals: number;
    blocks: number;
    assists_turnover_ratio: number;
    personal_fouls: number;
    offensive_fouls: number;
    tech_fouls: number;
    flagrant_fouls: number;
    pls_min: number;
    points: number;
    effective_fg_pct: number;
    efficiency: number;
    efficiency_game_score: number;
    points_in_paint: number;
    points_in_paint_att: number;
    points_in_paint_made: number;
    points_in_paint_pct: number;
    true_shooting_att: number;
    true_shooting_pct: number;
    fouls_drawn: number;
    points_off_turnovers: number;
    second_chance_pts: number;
    fast_break_pts: number;
    fast_break_att: number;
    fast_break_made: number;
    fast_break_pct: number;
    minus: number;
    plus: number;
    defensive_rebounds_pct: number;
    offensive_rebounds_pct: number;
    rebounds_pct: number;
    steals_pct: number;
    turnovers_pct: number;
    second_chance_att: number;
    second_chance_made: number;
    second_chance_pct: number;
    defensive_rating?: number;
    offensive_rating?: number;
}

export interface GameSummaryPlayersStatistics {
    minutes: string;
    field_goals_made: number;
    field_goals_att: number;
    field_goals_pct: number;
    three_points_made: number;
    three_points_att: number;
    three_points_pct: number;
    two_points_made: number;
    two_points_att: number;
    two_points_pct: number;
    blocked_att: number;
    free_throws_made: number;
    free_throws_att: number;
    free_throws_pct: number;
    offensive_rebounds: number;
    defensive_rebounds: number;
    rebounds: number;
    assists: number;
    turnovers: number;
    steals: number;
    blocks: number;
    assists_turnover_ratio: number;
    personal_fouls: number;
    tech_fouls: number;
    flagrant_fouls: number;
    pls_min: number;
    points: number;
    double_double: boolean;
    triple_double: boolean;
    effective_fg_pct: number;
    efficiency: number;
    efficiency_game_score: number;
    points_in_paint: number;
    points_in_paint_att: number;
    points_in_paint_made: number;
    points_in_paint_pct: number;
    true_shooting_att: number;
    true_shooting_pct: number;
    fouls_drawn: number;
    offensive_fouls: number;
    points_off_turnovers: number;
    second_chance_pts: number;
    fast_break_pts: number;
    fast_break_att: number;
    fast_break_made: number;
    fast_break_pct: number;
    coach_ejections: number;
    minus: number;
    plus: number;
    defensive_rebounds_pct: number;
    offensive_rebounds_pct: number;
    rebounds_pct: number;
    steals_pct: number;
    turnovers_pct: number;
    coach_tech_fouls: number;
    second_chance_att: number;
    second_chance_made: number;
    second_chance_pct: number;
    periods: Period[];
    defensive_rating?: number;
    offensive_rating?: number;
}

export interface GameSummaryPlayer {
    full_name: string;
    jersey_number: string;
    id: string;
    first_name: string;
    last_name: string;
    position: string;
    primary_position: string;
    active?: boolean;
    not_playing_reason?: string;
    on_court: boolean;
    sr_id: string;
    reference: string;
    statistics: GameSummaryPlayersStatistics;
    played?: boolean;
    starter?: boolean;
    not_playing_description?: string;
    name_suffix?: string;
}